f-feature-management-vue@v0.1.0 - use vue-demi to allow use in either Vue2 or Vue3
