+++ packages/services/f-feature-management-vue/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@justeat/f-feature-management-vue3",
+  "name": "@justeat/f-feature-management-vue",
   "version": "0.1.0",
   "description": "Service for querying feature values provided by Feature Management, made reactive for Vue3.",
-  "main": "dist/f-feature-management-vue3.umd.js",
-  "module": "dist/f-feature-management-vue3.es.js",
+  "main": "dist/f-feature-management-vue.umd.js",
+  "module": "dist/f-feature-management-vue.es.js",
   "source": "src/index.js",
   "files": [
     "dist"
@@ -36,18 +36,20 @@
   "devDependencies": {
     "@babel/core": "7.14.8",
     "@babel/preset-env": "7.14.9",
-    "@vue/cli-plugin-eslint": "3.9.2",
+    "@vue/cli-plugin-eslint": "3.9.2",    
     "babel-loader": "8.1.0",
     "eslint-plugin-json-schema-validator": "1.0.48",
     "eslint-plugin-jsonc": "1.3.1",
     "jest": "25.4.0",
     "jest-teamcity": "1.7.0",
     "vite": "2.4.4",
-    "@vue/reactivity": "3.2.11"
+    "vue": "^3.0.0"
   },
   "peerDependencies": {
     "@justeat/f-feature-management": "0.4.0",
-    "@vue/reactivity": "3.2.11"
+    "vue": "vue@^3.0.0-0 || ^2.0.0"
   },
-  "dependencies": { }
+  "dependencies": {
+    "vue-demi": "0.11.4"
+  }
 }
