+++ packages/services/f-feature-management-vue/test/createFeatureManagementForVue.test.js
@@ -1,4 +1,4 @@
-import createFeatureManagementForVue3 from '../src/createFeatureManagementForVue3';
+import createFeatureManagementForVue from '../src/createFeatureManagementForVue';
 import createFeatureManagement from '@justeat/f-feature-management';
 import { reactive } from '@vue/reactivity';
 
@@ -14,10 +14,11 @@ jest.mock('@justeat/f-feature-management', () => {
 });
 
 describe('When we get a mock Feature Management object', () => {
+    //NOTE: this would fail if our devDependency was Vue2
     it('should be a reactive proxy', async () => {          
 
         //Act
-        const fm = await createFeatureManagementForVue3({});
+        const fm = await createFeatureManagementForVue({});
 
         //Assert
         //Vue3 tracks reactive proxies, so calling reactive on the same object twice will yield the same return object
@@ -28,7 +29,7 @@ describe('When we get a mock Feature Management object', () => {
     it('should have a property called reactive with a getValue function on it', async () => {          
 
         //Act
-        const fm = await createFeatureManagementForVue3({});
+        const fm = await createFeatureManagementForVue({});
 
         //Assert
         expect(fm.reactive).toBeInstanceOf(Object);
@@ -42,8 +43,8 @@ describe('When we get a mock Feature Management object', () => {
         mockFm.getValue.mockReturnValue('return value');
         
         //Act
-        const fm = await createFeatureManagementForVue3({});
-        const val = fm.getValue('parameter');
+        const fm = await createFeatureManagementForVue({});
+        const val = fm.reactive.getValue('parameter');
 
         //Assert
         expect(mockFm.getValue).toHaveBeenCalledWith('parameter');
