--- packages/services/f-feature-management-vue3/vite.config.js
@@ -1,11 +1,0 @@
-import path from 'path';
-
-export default {
-    build: {
-        target: 'es2015',
-        lib: {
-            entry: path.resolve(__dirname, 'src/index.js'),
-            name: 'f-feature-management-vue3'
-        }
-    }
-};
